Move inverse and transpose tests to their own files.
