+++ sympy/matrices/expressions/tests/test_inverse.py
@@ -1,0 +1,27 @@
+from sympy.core import symbols, S
+from sympy.functions import adjoint, conjugate, transpose
+from sympy.matrices.expressions import MatrixSymbol, Inverse
+from sympy.matrices import eye, Identity, Matrix, ShapeError
+from sympy.utilities.pytest import raises
+
+n, m, l = symbols('n m l', integer=True)
+A = MatrixSymbol('A', n, m)
+B = MatrixSymbol('B', m, l)
+C = MatrixSymbol('C', n, n)
+D = MatrixSymbol('D', n, n)
+E = MatrixSymbol('E', m, n)
+
+
+def test_inverse():
+    raises(ShapeError, lambda: Inverse(A))
+    assert C.inverse().inverse() == C
+
+    assert C.inverse()*C == Identity(C.rows)
+
+    assert Identity(n).inverse() == Identity(n)
+    assert (3*Identity(n)).inverse() == Identity(n)/3
+
+    # Simplifies Muls if possible (i.e. submatrices are square)
+    assert (C*D).inverse() == D.I*C.I
+    # But still works when not possible
+    assert isinstance((A*E).inverse(), Inverse)
